Added a wrapper in MainClass for M operations
Added a wrapper methos executeOperationOnM in the main class for M
Operations.

This method is called from the scriptanalyzer for an M Query.
I gets the in memory and after copy contents for the M Query.
displays these results
Checks if a delete on that table exists. if it does not then calls the
columns store method for M Query.